pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arp-jarvis-node-registry"
version = "0.3.8"
description = "JARVIS Node Registry implementation for the ARP Standard (spec/v1)."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "Agent Runtime Protocol" }]
dependencies = [
  "arp-standard-client==0.3.7",
  "arp-standard-server==0.3.7",
  "arp-standard-model==0.3.7",
  "arp-jarvis-atomic-nodes[metadata]==0.3.8",
  "uvicorn>=0.29.0",
]

[project.optional-dependencies]
dev = ["pyright>=1.1.0", "pytest>=7", "pytest-cov>=4"]

[project.scripts]
arp-jarvis-node-registry = "jarvis_node_registry.__main__:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
